requirements.txt
Flask==3.1.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
authlib==1.3.0
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.0.1
